requirements.txt
transformers
torch
accelerate
torchvision
pillow
numpy
paddlepaddle
paddleocr
qwen-vl-utils
